+++ src/views/products/components/ProductsTable.scss
