+++ src/views/products/components/ProductsTable.tsx
@@ -1,4 +1,4 @@
-import { Product } from './types';
+import { Product } from '../types';
 import {
   TableContainer,
   Paper,
